+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# TBD
+
++ Add "winrm config" and "winrm run" operations
++ Add "winrm copy" and "winrm fetch" operations
+
 # v0.8.9
 
 + Add "support" operation.